Split the layout selectors between graph-view and resource-view
