+++ client/app/scripts/selectors/graph-view/graph.js
@@ -2,11 +2,11 @@ import debug from 'debug';
 import { createSelector, createStructuredSelector } from 'reselect';
 import { Map as makeMap } from 'immutable';
 
-import { initEdgesFromNodes } from '../utils/layouter-utils';
-import { viewportWidthSelector, viewportHeightSelector } from './canvas-viewport';
-import { activeTopologyOptionsSelector } from './topology';
-import { shownNodesSelector } from './node-filters';
-import { doLayout } from '../charts/nodes-layout';
+import { initEdgesFromNodes } from '../../utils/layouter-utils';
+import { viewportWidthSelector, viewportHeightSelector } from './viewport';
+import { activeTopologyOptionsSelector } from '../topology';
+import { shownNodesSelector } from '../node-filters';
+import { doLayout } from '../../charts/nodes-layout';
 import timer from '../utils/timer-utils';
 
 const log = debug('scope:nodes-chart');
